Update Inventory Controller v.5
Installation of the removeNewline function, bugfixing of the new Below and Above hidden text areas (implemented due to TTsimulator deleting surplus card text when edited).
This version is unstable. I plan to depreciate the plainReplace function in favor of a length-based replacement system. This, in conjunction with the removeNewline function, entails a large rework of the way scroll up and scroll down handle edge cases.