Update and rename Inventory Controller v.5 to Inventory Controller v.6
I'll need to clean out debug functions and whatnot, but basic functionality for the controller is all intact (deals with dynamic text now). I haven't been able to break it yet, but I'll need to do more testing to really be sure.

Upcoming, I'll make an initialization part for the player cards in case onSave doesn't have the values.