Create D&D 5e Character Sheet Source
This is the character sheet brain, or "source", from which all of the positional data for character sheets is read (all of the offset tables have been moved here). Additionally, the calculation function for character sheet automation is also stored here.

This will allow for any custom sheet to be read and used by the table once it is loaded onto the special platform.