Create Blank Character Sheet
I think I accidently deleted this before.

This is the "barebones" character sheet. It needs a diffuse and I'm not sure how the game will react without any DM_Visible_Stats to back up the names, but is otherwise the starting point for someone who wants to make their own custom sheet.